perf(src): switch mapHolderRef from props to context
* Improved performance of rendering many markers/components on the map
* Closes #135
* Closes #210
* Closes #216

BREAKING CHANGE: If you're just using the library and didn't make custom components before, feel free to ignore this implementation changes.

Now, mapHolderRef for each component are now passed in via context. This doesn't affect all components interface in general. But if you do custom components before and relies on the implementation of react-google-maps, you should be aware of this and make some changes.